requirements.txt
google-genai
PyYAML
openai
